+++ windy-common/src/main/java/com/zj/common/exception/GlobalExceptionHandler.java
@@ -19,25 +19,23 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-  public static final String LEFT_STRING = "[";
-  public static final String RIGHT_STRING = "] ";
+  public static final String LEFT_STRING = "【";
+  public static final String RIGHT_STRING = "】 ";
 
   @ExceptionHandler(Exception.class)
   @ResponseBody
   public ResponseEntity<ResponseMeta> handleException(Exception e) {
-    ErrorCode errorCode = ErrorCode.ERROR;
-    errorCode.bindMessage(e.getMessage());
-
+    String errorCode = ErrorCode.ERROR.getCode();
+    String message = e.getMessage();
     if (e instanceof MethodArgumentNotValidException) {
-      errorCode = resetParamErrorMsg(e);
+      log.warn("param validate error", e);
+      message = resetParamErrorMsg(e);
     }
-
-    log.warn("param validate error", e);
-    return new ResponseEntity<ResponseMeta>(new ResponseMeta(errorCode),
+    return new ResponseEntity<>(new ResponseMeta(errorCode, message, null),
         HttpStatus.INTERNAL_SERVER_ERROR);
   }
 
-  private ErrorCode resetParamErrorMsg(Exception e) {
+  private String resetParamErrorMsg(Exception e) {
     List<FieldError> errors = ((MethodArgumentNotValidException) e)
         .getBindingResult().getFieldErrors();
     StringBuilder stringBuilder = new StringBuilder();
@@ -45,17 +43,14 @@ public class GlobalExceptionHandler {
       stringBuilder.append(LEFT_STRING).append(fieldError.getField()).append(RIGHT_STRING)
           .append(fieldError.getDefaultMessage());
     });
-
-    ErrorCode errorCode = ErrorCode.PARAM_VALIDATE_ERROR;
-    errorCode.bindMessage(stringBuilder.toString());
-    return errorCode;
+    return stringBuilder.toString();
   }
 
   @ExceptionHandler(CommonException.class)
   @ResponseBody
   public ResponseEntity<ResponseMeta> handleBusinessException(CommonException commonException) {
     log.warn("common error", commonException);
-    return new ResponseEntity<ResponseMeta>(new ResponseMeta(commonException.getErrorCode()),
+    return new ResponseEntity<>(new ResponseMeta(commonException.getErrorCode()),
         commonException.getErrorCode().getHttpStatus());
   }
 }
