--- windy-pipeline/src/main/java/com/zj/pipeline/mapper/MicroserviceMapper.java
@@ -1,7 +1,0 @@
-package com.zj.pipeline.mapper;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.zj.pipeline.entity.po.Microservice;
-
-public interface MicroserviceMapper extends BaseMapper<Microservice> {
-}
