+++ windy-pipeline/src/main/java/com/zj/pipeline/service/CodeChangeService.java
@@ -8,9 +8,10 @@ import com.zj.pipeline.entity.dto.RelationDemandBug;
 import com.zj.pipeline.entity.po.CodeChange;
 import com.zj.common.exception.ApiException;
 import com.zj.common.exception.ErrorCode;
-import com.zj.pipeline.entity.po.Microservice;
+import com.zj.service.entity.po.Microservice;
 import com.zj.pipeline.git.IRepositoryBranch;
 import com.zj.pipeline.mapper.CodeChangeMapper;
+import com.zj.service.service.MicroserviceService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
