+++ windy-pipeline/src/main/java/com/zj/pipeline/service/GitBindService.java
@@ -5,16 +5,17 @@ import com.alibaba.fastjson.JSONObject;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.zj.pipeline.entity.dto.GitBindDto;
+import com.zj.pipeline.entity.dto.MicroserviceDto;
 import com.zj.pipeline.entity.dto.PipelineDTO;
 import com.zj.pipeline.entity.enums.PipelineExecuteType;
-import com.zj.pipeline.entity.enums.PipelineType;
 import com.zj.pipeline.entity.po.GitBind;
 import com.zj.common.exception.ApiException;
 import com.zj.common.exception.ErrorCode;
-import com.zj.pipeline.entity.po.Microservice;
+import com.zj.pipeline.git.IRepositoryBranch;
+import com.zj.service.entity.po.Microservice;
 import com.zj.pipeline.entity.po.Pipeline;
-import com.zj.pipeline.executer.vo.ExecuteType;
 import com.zj.pipeline.mapper.GitBindMapper;
+import com.zj.service.service.MicroserviceService;
 import java.util.Optional;
 import java.util.concurrent.ExecutorService;
 import lombok.extern.slf4j.Slf4j;
@@ -47,6 +48,9 @@ public class GitBindService extends ServiceImpl<GitBindMapper, GitBind> {
   @Autowired
   @Qualifier("webHookExecutorPool")
   private ExecutorService executorService;
+
+  @Autowired
+  private IRepositoryBranch repositoryBranch;
 
   public String createGitBind(GitBindDto gitBindDto) {
     List<GitBindDto> bindDtoList = listGitBinds(gitBindDto.getPipelineId());
@@ -149,5 +153,10 @@ public class GitBindService extends ServiceImpl<GitBindMapper, GitBind> {
     String ref = data.getString("ref");
     int index = ref.lastIndexOf("/");
     return ref.substring(index + 1);
+  }
+
+  public List<String> getServiceBranch(String serviceId) {
+    Microservice serviceDetail = microserviceService.getServiceDetail(serviceId);
+    return repositoryBranch.listBranch(serviceDetail.getServiceName());
   }
 }
